refactor listener
