+++ lib/listen/change.rb
@@ -20,15 +20,8 @@ module Listen
         return
       end
 
-      if options[:silence]
-        _log :debug, "recording: #{type}:#{path} (#{options.inspect})"
-      else
-        if change
-          _log :debug, "#{change}: #{type}:#{path} (#{options.inspect})"
-        else
-          _log :debug, "unknown: #{type}:#{path} (#{options.inspect})"
-        end
-      end
+      log_details = options[:silence] && 'recording' || change || 'unknown'
+      _log :debug, "#{log_details}: #{type}:#{path} (#{options.inspect})"
 
       if change
         listener.queue(type, change, path, cookie ? { cookie: cookie } : {})
